requirements.txt
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.1
gcloud==0.18.3
google-api-core==2.10.0
google-auth==2.11.0
google-cloud-core==2.3.2
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
httplib2==0.20.4
idna==3.3
numpy==1.23.2
oauth2client==4.1.3
pandas==1.4.4
protobuf==4.21.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
rsa==4.9
six==1.16.0
urllib3==1.26.12
